+++ Backend/aptitudeproject/src/main/java/com/aptitude/repository/UserAnswerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.aptitude.entity.UserAnswer;
 
 @Repository
-public interface UserAnswerRepository extends JpaRepository<UserAnswer, Integer> {
+public interface UserAnswerRepository extends JpaRepository<UserAnswer, Long> {
 
 }
